+++ pages/Hello.py
@@ -12,7 +12,7 @@ def main():
     st.page_link("pages/Info.py", label="Go to the Info page", icon="📈")
     st.page_link("pages/Compare.py", label="Go to the Compare page", icon="🔀")
     st.page_link("pages/Prediction.py", label="Go to the Prediction page", icon="🔮")
-    st.page_link("pages/News2.py", label="Go to the News page", icon="📰")
+    st.page_link("pages/News.py", label="Go to the News page", icon="📰")
